+++ tests/test_builders.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
-from sphinxcontrib.httpexample.parsers import parse_request
-from sphinxcontrib.httpexample.builders import build_curl_command
-from sphinxcontrib.httpexample.builders import build_wget_command
-from sphinxcontrib.httpexample.builders import build_httpie_command
-from sphinxcontrib.httpexample.builders import build_requests_command
-from sphinxcontrib.httpexample.builders import is_json
+from sphinxcontrib.httpexmpl.parsers import parse_request
+from sphinxcontrib.httpexmpl.builders import build_curl_command
+from sphinxcontrib.httpexmpl.builders import build_wget_command
+from sphinxcontrib.httpexmpl.builders import build_httpie_command
+from sphinxcontrib.httpexmpl.builders import build_requests_command
+from sphinxcontrib.httpexmpl.builders import is_json
 
 from .test_fixtures import FIXTURE_001_REQUEST
 from .test_fixtures import FIXTURE_002_REQUEST