+++ tests/test_utils.py
@@ -3,7 +3,7 @@ import json
 import os
 import pytest
 
-from sphinxcontrib.httpexample import utils
+from sphinxcontrib.httpexmpl import utils
 
 
 def test_merge_dicts():
@@ -15,7 +15,7 @@ def test_resolve_path():
     cwd = os.path.dirname(__file__)
     base = os.path.basename(__file__)
     assert utils.resolve_path(base, cwd) == __file__
-    assert utils.resolve_path('sphinxcontrib.httpexample:utils.py') == utils.__file__  # noqa
+    assert utils.resolve_path('sphinxcontrib.httpexmpl:utils.py') == utils.__file__  # noqa
     assert utils.resolve_path('bar', 'non-existing') == 'bar'